# perf: optimize test suite execution time (6.5x speedup)