--- a/docs/handoff/2025-06-08-2.md
+++ b/docs/handoff/2025-06-08-2.md
@@ -1,0 +1,61 @@
+# Handoff Document - 2025-06-08 (#2)
+
+## Session Summary
+
+Optimized test suite performance achieving 6.5x speedup (36s → 5.6s) through targeted optimizations and parallel execution.
+
+## Key Changes
+
+### Test Optimizations
+1. **SQLite Backup Tests** (`tests/luca_core_pkgtests/test_sqlite_store.py`)
+   - Replaced `asyncio.sleep(1.5)` with direct backup method calls
+   - Changed backup interval from 1s to 3600s in test fixture
+   - Saved 3 seconds total
+
+2. **Sandbox Timeout Test** (`tests/test_sandbox_manager.py`)
+   - Reduced timeout from 1s to 0.1s for timeout validation
+   - Test still validates timeout behavior correctly
+   - Saved 0.9 seconds
+
+3. **Parallel Test Execution**
+   - Added pytest-xdist==3.5.0 to requirements-dev.txt
+   - Created `make test-fast` target for parallel tests
+   - Updated safety-check.sh to auto-detect and use parallel testing
+   - Achieved 557% CPU utilization (multi-core)
+
+### Documentation
+- Created comprehensive research document: `docs/research/test-suite-optimization.md`
+- Updated Makefile help text to include test-fast target
+- Fixed import issues caught by linting
+
+## Performance Results
+
+```
+Sequential: 36 seconds → 33 seconds (with optimizations)
+Parallel:   36 seconds → 5.6 seconds (6.5x speedup)
+Pre-push:   90+ seconds → ~20 seconds
+```
+
+## Technical Details
+
+- Used pytest-xdist 3.5.0 (not 3.6.0 which has pytest-cov compatibility issues)
+- Coverage collection works seamlessly with parallel execution
+- All tests pass with maintained coverage (96%)
+
+## Next Steps
+
+1. Monitor CI performance with these optimizations
+2. Consider additional optimizations:
+   - Mock subprocess calls in watchdog tests
+   - Use in-memory SQLite for test databases
+   - Implement test result caching
+
+## Environment
+- Branch: `feature/test-optimization`
+- Python: 3.13.3
+- Working directory: `/Users/dustinkirby/Documents/GitHub/luca-dev-assistant`
+
+## Notes
+- Pre-commit hooks now complete in reasonable time
+- Developer experience significantly improved
+- No functionality changes, only test performance improvements