--- a/docs/handoff/2025-06-08-2.md
+++ b/docs/handoff/2025-06-08-2.md
@@ -4,7 +4,7 @@
 
 Optimized test suite performance achieving 6.5x speedup (36s → 5.6s) through targeted optimizations and parallel execution.
 
-## Key Changes
+## Work Completed
 
 ### Test Optimizations
 1. **SQLite Backup Tests** (`tests/luca_core_pkgtests/test_sqlite_store.py`)
@@ -35,6 +35,15 @@ Sequential: 36 seconds → 33 seconds (with optimizations)
 Parallel:   36 seconds → 5.6 seconds (6.5x speedup)
 Pre-push:   90+ seconds → ~20 seconds
 ```
+
+## Current State
+
+- **Branch**: `feature/test-optimization` (ready for PR)
+- **Tests**: All passing with 96.11% coverage
+- **Performance**: Tests run in 5.6s (parallel) vs 36s (sequential)
+- **Pre-push hooks**: Now complete in ~20s (was 90+s)
+- **CI**: Expected to see significant improvement
+- **Code Quality**: All linting, formatting, and security checks pass
 
 ## Technical Details