--- a/docs/handoff/2025-06-08-2.md
+++ b/docs/handoff/2025-06-08-2.md
@@ -6,24 +6,25 @@ Optimized test suite performance achieving 6.5x speedup (36s → 5.6s) through t
 
 ## Work Completed
 
-### Test Optimizations
-1. **SQLite Backup Tests** (`tests/luca_core_pkgtests/test_sqlite_store.py`)
-   - Replaced `asyncio.sleep(1.5)` with direct backup method calls
-   - Changed backup interval from 1s to 3600s in test fixture
-   - Saved 3 seconds total
+### 1. Test Optimizations
+- **SQLite Backup Tests** (`tests/luca_core_pkgtests/test_sqlite_store.py`)
+  - Replaced `asyncio.sleep(1.5)` with direct backup method calls
+  - Changed backup interval from 1s to 3600s in test fixture
+  - Saved 3 seconds total
 
-2. **Sandbox Timeout Test** (`tests/test_sandbox_manager.py`)
-   - Reduced timeout from 1s to 0.1s for timeout validation
-   - Test still validates timeout behavior correctly
-   - Saved 0.9 seconds
+### 2. Sandbox Timeout Test
+- **File**: `tests/test_sandbox_manager.py`
+- Reduced timeout from 1s to 0.1s for timeout validation
+- Test still validates timeout behavior correctly
+- Saved 0.9 seconds
 
-3. **Parallel Test Execution**
-   - Added pytest-xdist==3.5.0 to requirements-dev.txt
-   - Created `make test-fast` target for parallel tests
-   - Updated safety-check.sh to auto-detect and use parallel testing
-   - Achieved 557% CPU utilization (multi-core)
+### 3. Parallel Test Execution
+- Added pytest-xdist==3.5.0 to requirements-dev.txt
+- Created `make test-fast` target for parallel tests
+- Updated safety-check.sh to auto-detect and use parallel testing
+- Achieved 557% CPU utilization (multi-core)
 
-### Documentation
+### 4. Documentation
 - Created comprehensive research document: `docs/research/test-suite-optimization.md`
 - Updated Makefile help text to include test-fast target
 - Fixed import issues caught by linting