--- a/docs/handoff/2025-06-08-2.md
+++ b/docs/handoff/2025-06-08-2.md
@@ -39,12 +39,12 @@ Pre-push:   90+ seconds → ~20 seconds
 
 ## Current State
 
-- **Branch**: `feature/test-optimization` (ready for PR)
-- **Tests**: All passing with 96.11% coverage
-- **Performance**: Tests run in 5.6s (parallel) vs 36s (sequential)
-- **Pre-push hooks**: Now complete in ~20s (was 90+s)
-- **CI**: Expected to see significant improvement
-- **Code Quality**: All linting, formatting, and security checks pass
+- Branch: `feature/test-optimization` (ready for PR)
+- Tests: All passing with 96.11% coverage
+- Performance: Tests run in 5.6s (parallel) vs 36s (sequential)
+- Pre-push hooks: Now complete in ~20s (was 90+s)
+- CI: Expected to see significant improvement
+- Code Quality: All linting, formatting, and security checks pass
 
 ## Technical Details