--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -12,6 +12,7 @@ types-PyYAML==6.0.12.20250516
 pytest-asyncio==0.26.0  # For async test support
 pytest-mock==3.14.1     # For mocking in tests
 pytest-cov==6.1.1       # For coverage configuration
+pytest-xdist==3.5.0     # For parallel test execution
 
 # Security
 bandit==1.8.3